Add a sentry example.
The main idea behind sentries (sealed entries) is that they protect the
function pointers from being modified, read to/from. This allows emulation
of most sealed capability modes. This example shows some of the use cases
of sentry capabilities that I could think of. It is important to note that every
function pointer on CHERI is using sentry capabilities.