Add general example regarding bounds
This example wants to clarify the difference between
setting the bounds of a capable pointer explicitly
and dereferencing a pointer outside the bounds on
two architectures: "riscv64" and "Morello" (ARMv8).

CHERI limits the range of addresses that may be dereferenced
and, in this example, reading the value outside the range will
give an "In-address space security exception" on "Morello".
However, on "riscv64", we can make the program crash by
explicitly setting the bounds outside the allowed range.