Remove redundant code
The example showing compartment switching with sentries had redundant
code. We do not need to explicitly grant execution permission to the
function wrapped by the sentry, as this exists.

The issue before was casting the wrapped function pointer to a `void*`,
rather than directly to `void* __capability`, which removed the
necessary permissions.