Add compartment example
This example showcases two particular actions to help compartmentalization:
     - reducing the bounds of the PCC (shared.s:82)
     - calling a sentry from within a bounded compartment (shared.s:114)
In addition to the above, we note that an implicit PCC unbounding is
performed when a sentry is called in hybrid mode. The PCC is then **not**
implicitly rebound. This example uses a trampoline function `comp_fun_tr`
which does this rebounding.