Add two compartment examples
These examples aim to do two things:
* restrict the program capability counter (`PCC`)
* call a function outside PCC bounds via a sealed capability entry
  (sentry)

Example `restrict_pcc` shows that we successfully restrict the `PCC`,
and that calling a function outside the `PCC` bounds fails, with an
expected `In-address space security exception`.

Example `call_sentry` shows how to call a function outside of restricted
`PCC` bounds via a sentry, as well as one (not fully secure) choice of
how to transition between a compartment (i.e., restricted permissions)
and a state with full permissions.