Improve compartment security
In this example, we tighten security even further. We store the DDC/PCC
of `switch_compartment` within its memory region. Then, for each
compartment, we create a copy of a capability pointing to this DDC/PCC
pair, sealed with `lpb`. This allows compartments to call
`switch_compartment` without having access to any information pertaining
to it.

We provide an additional example (`shared_try_deref.S`), where a
malicious compartment attempts to dereference its local copy, to gain
access to `switch_compartment`.