Fix `switch_compartment` PCC bounds
A hardcoded value of the bounds over `switch_compartment` meant that the
bounds were much wider than expected. This meant two things:
- the `clean` function called within `switch_compartment` was reachable
  and executable;
- when deriving the `clr` within `switch_compartment`, we would derive
  with the larger bound, and return to a compartment with much more
  executable permissions than intended.

We now tightly bind the PCC to `switch_compartment`. This means moving
`clean` within these bounds (a better way would be to create
capabilities allowing this function to be called, and provide local
copies to each compartment, but that is beyond the scope of this
example, and an exercise in engineering), and not deriving `clr` within
`switch_compartment`, but just retaining the provided `clr`.

Additional small cleanups, such as code-base splitting, and
comment-fixing.