Add an example that reads a capability from a file.
According to the CHERI C/C++ programming guide:

> In the CheriABI process environment, only untagged data (not tagged
> pointers) may be written to or read from files. If a region of memory
> containing valid pointers is written to a file, and then read back, the
> pointers in that region24 will no longer be valid.

However, as this example demonstrates, this behaviour appears to be
broken, as it is possible to read a valid capability from a file.

This example writes a capability to a temporary file and then
successfully retrieves the original (valid) capability, which should not
be possible according to the CheriABI.

The problem still reproduces if we simplify the example further, by
removing the part where the capability is written to disk and
hard-coding the file from which to read the capability (the file could
contain `ff5c fff7 ffff 0000 ff5c 7f60 4000 dc5d` for example).