A couple of examples which fail to leak capabilities using pipes.
The first example (`read-cap-from-pipe`) consists of two sibling
processes connected to a pipe:
   * a writer process that writes a capability to the pipe
     (`write_to_pipe`), and
   * a reader process that reads the capability written by the writer
     process to the pipe (`read_from_pipe`). The reader process reads
     an invalid capability

The second example (`read-cap-from-pipe-sh`) is similar, except it
relies on the shell to connect the two processes to a pipe.

In both examples, the reader process fails to read a valid capability
from the pipe (as expected).