Added test of the compare_platforms/compare_platforms_read_only example to the test script tests/run_tests.sh, to make sure the example executes as expected (it should fail).
The example has a struct, 'review', with three members:
- 'realname', a char* representing the reviewer's name
- 'publicreview', a char* representing a review of a submitted paper
- 'permissions', a char* that is initialized to just 'r'

There are two sizes passed to malloc(): 0x20 and 0x1001.
- 'realname' and 'permissions' are allocated the smaller size before
being initialized.
- 'publicreview' is allocated the larger size, and then initialized to
a few sentences rejecting an imagined paper submission.

The relative positions of the three pointers are checked, by a series
of asserts that their respective allocations should not overlap.

Then, another char* called 'newpublicreview' is created, containing
a much more positive review and accepting the paper.

Next, in a CHERI environment only, capabilities are used to set the
members of the 'review' struct to read-only.

At this point there is one execution path, exclusive to CHERI, that
cannot be tested by the test script. If the command-line flag
'-overwrite' is passed in, the program will attempt to change
'publicreview' to 'newpublicreview', which will SIGPROT, because
all the struct's members have read-only permissions enforced by
capabilities. The test script cannot currently handle arguments passed
to the tests themselves, which is why this execution path cannot
currently be tested.

The other execution path, which can be tested by the current script,
is followed both in a CHERI environment where the '-overwrite' flag
is not passed in, and in a non-CHERI environment.

In this second execution path, the program will try to overflow the
bounds of 'realname' in order to add the character 'w' to 'permissions',
the presence of which character is interpreted as conferring write
permissions on all of the struct's members.

The program first attempts to overflow 'realname' by one, then by
multiple characters.
- In a CHERI environment, this will cause a SIGPROT, though different
implementations may SIGPROT at different points in the process.
- In a non-CHERI environment, the overflow is expected to succeed,
acquire write permissions, and rewrite the review such that the
rejected paper appears to have been accepted by the reviewer.

Finally, the pointer allocations are freed. This code will only be
reached in a non-CHERI environment, and, depending on the environment,
may crash.