Add `compartment_alloc` "side compartment" example
Adds an example allocator which manages "side compartments", which are
introduced in this example's README. There are two functions here which
are relevant to creating compartments and allocating memory within them.

1. `init_compartment` `mmap`s an area of memory, creates a "compartment"
   (see `README` for a high-level definition, or `compartment_alloc.h`
   for the definition of the struct itself) and returns a sealed
   capability which can be used to identify a compartment when
   allocating memory.
2. `malloc_compartment` allocates memory within a compartment's `mmap`'d
   buffer. It accepts two arguments: a number of bytes to allocate, and
   a capability identifying a compartment (i.e. one returned by
   `init_compartment`). It returns a capability which points to memory
   within the compartment, assuming it's able to allocate (errors such
   as running our of space cause the program to exit with RC 1, because
   this is an example and it doesn't have to handle things particularly
   gracefully).

There's also a function to free a compartment — `free_compartment` —
which takes a compartment's identifying capability and frees the mmap'd
buffer associated with it. This is naive and not a complete
implementation; capabilities pointing to somewhere in that now-free
buffer are still valid capabilities, but will be unsafe to use.