Fix CI not failing
Unfortunately, when we moved calling the underlying CHERI instance via
`script` (in #93), we omitted to ensure that the return code for the
underlying tests was properly returned to the testing script it's being
called from. This means we had some underlying silent failures in our
CI. We attempt to fix them here, and to ensure future CI checks fail
appropriately.